fix race condition
if we insert an entry into the cachemap before updating the policy, a
reader can start reading the data and try to update the policy
metadata before the writer thread inserts an entry into the eviction
policy metadata. then the reader thread will encounter a NULL when
trying to find the chunk in the LRU queue. this issue didn't pop up
with zone promotion because a NULL in the LRU queue in the zone
promotion policy was valid as the entry could have been evicted by the
eviction thread. so essentially this also occured with zone promotion,
but it was tolerated. at worst, it would have meant a missed update
from a read.

by moving the cachemap update after the eviction policy update, no
reader threads will be able to read before the writer thread finishes
creating the entry in the eviction policy, thus avoiding the issue.

we still need to only check for node and not node->data, because it
can still potentially be NULL for the same reason that zone entries
could potentially be NULL in the promotional policy: while a reader
thread was reading a chunk, the eviction thread could have popped that
chunk off for eviction, and so when the reader thread tries to update
the policy, it would find that the chunk entry has already been
deleted.